Заэкстендил функции экстра фильмов
